+++ mpv.net/Misc/Program.cs
@@ -1,5 +1,10 @@
 ﻿using System;
 using System.Windows.Forms;
+using System.Linq;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading;
+using System.Diagnostics;
 
 namespace mpvnet
 {
@@ -8,15 +13,43 @@ namespace mpvnet
         [STAThread]
         static void Main()
         {
+            Mutex mutex = new Mutex(true, "mpvnetProcessInstance", out bool isFirst);
+
             try
             {
-                string[] args = Environment.GetCommandLineArgs();
+                string[] args = Environment.GetCommandLineArgs().Skip(1).ToArray();
 
-                if (args.Length == 3 && args[1] == "--reg-file-assoc")
+                if (args.Length == 2 && args[0] == "--reg-file-assoc")
                 {
-                    if (args[2] == "audio") FileAssociation.Register(App.AudioTypes);
-                    if (args[2] == "video") FileAssociation.Register(App.VideoTypes);
-                    if (args[2] == "unreg") FileAssociation.Unregister();
+                    if (args[1] == "audio") FileAssociation.Register(App.AudioTypes);
+                    if (args[1] == "video") FileAssociation.Register(App.VideoTypes);
+                    if (args[1] == "unreg") FileAssociation.Unregister();
+                    return;
+                }
+
+                App.Init();
+
+                if ((App.ProcessInstance == "single" || App.ProcessInstance == "queue") && !isFirst)
+                {
+                    List<string> files = new List<string>();
+
+                    foreach (string arg in args)
+                        if (!arg.StartsWith("--") && (File.Exists(arg) || arg == "-" || arg.StartsWith("http")))
+                            files.Add(arg);
+
+                    if (files.Count > 0)
+                        RegistryHelp.SetObject(App.RegPath, "ShellFiles", files.ToArray());
+
+                    RegistryHelp.SetObject(App.RegPath, "ProcessInstanceMode", App.ProcessInstance);
+
+                    foreach(Process process in Process.GetProcessesByName("mpvnet"))
+                    {
+                        try {
+                            SingleProcess.AllowSetForegroundWindow(process.Id);
+                            Native.SendMessage(process.MainWindowHandle, SingleProcess.Message, IntPtr.Zero, IntPtr.Zero);
+                        } catch {}
+                    }
+
                     return;
                 }
 
@@ -27,6 +60,10 @@ namespace mpvnet
             catch (Exception ex)
             {
                 Msg.ShowException(ex);
+            }
+            finally
+            {
+                mutex.Dispose();
             }
         }
     }