+++ Changelog.md
@@ -5,6 +5,9 @@
 - the help and layout in the config editor was improved
 - clipboard monitoring for URLs can be disabled in the settings
 - internal refactoring that might break features, scripts or extensions
+- the context menu has a new feature: Open > Add files to playlist,
+  it appends files to the playlist [(Default binding)](https://github.com/stax76/mpv.net/blob/master/mpv.net/Resources/inputConf.txt#L34)
+- a new feature was added to force using a single mpv.net process instance
 
 ### 4.1